daemon/requirements.txt
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
urllib3==2.0.5 